# This is a combination of 7 commits. # The first commit's message is: Update ApplicationMaster.java # This is the 2nd commit message:
Update input_split_base.h
# This is the 3rd commit message:

Make concurrency g++-4.6 compatible

# This is the 4th commit message:

[DMLC-CORE] Support reset in InputSplit, remove Threaded Prefetcher of Input Split

# This is the 5th commit message:

fix lint
# This is the 6th commit message:

Specify `host` in container's env.

The current implementation gets ip/host via querying network adaptor info. This will raise issues when the ip/host got inside the container is not necessarily accessible from outside. To fix this, we can use the YARN node host of the container, which is guaranteed to be accessible from other containers and the YARN client node.

This commit just set the node host to the container's env. To use the node host, we can get the env from the application program executed on YARN. I will be committing that part of the change in dmlc/ps-lite.

# This is the 7th commit message:

add fault tolerance