data: Allow additional args in data parsers
This commit adds a new argument to the Parser::Create() function
and RowBlockIter::Create() function to add additional args/params
to the parser. These additional args are merged with the args
from the URI to get an effective list of args.
The args mentioned in the URI are considered as lower priority.