text_parser: Use the given nthread
Earlier, if a nthread was given to the Parser, the parser saves
that value and runs only those many parallel OMP loops. But the
calculation to find the blocks of the file were not using this
given nthread and always used the max number of threads. Which
results in only a part of the file actually being loaded
(given_nthread / max_nthread).

This uniformly uses the nthread_ value throughout.