parameter.h: Allow ranges and bracket less vector<int>s
vector<int> are essentially range of values and sometimes it is
easier to give a direct rang (Like 1-100) rather than typing the
whole vector (Like 1,2,3,4...). Hence allow ranges in this.

Also, as this is a specific case of parameters, allow bracket-less
vectors where "1,2,3" is directly picked up as the vector and does
not require brackets like "(1,2,3)".