Fix S3 pagination issue (#398)
When more than a 1000 files are found in an S3 bucket, S3 will paginate results
and multiple requests are necessary to retrieve them all (in which case the
response is said to be truncated). Previously, a check was done on
`isTruncated` and an exception raised if set to true. This commit fixes that
and retrieve all results.

Fix dmlc/dmlc-core#223